requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1
django-bootstrap5==24.2
django-extensions==3.2.3
django-filter==24.3
django-taggit==6.0.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.7
Pillow==10.1.0
psycopg==3.1.19
psycopg-binary==3.1.19
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
